Rename, tidy and add geo location
